fix(release): resolve PR review comments and centralize script logic into check_status.py
